requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
ciso8601==2.2.0
colorama==0.4.4
colorlog==6.6.0
dnspython==2.2.0
docopt==0.6.2
ed25519==1.5
frozenlist==1.3.0
hikari==2.0.0.dev106
hikari-lightbulb==2.2.0
hikari-miru==1.1.1
idna==3.3
lightbulb-ext-filament @ git+https://github.com/tandemdude/lightbulb-ext-filament@43351ab93b3c2ef1de47ce4539eaa20a5a580a85
multidict==6.0.2
nusex==1.2.4
pure25519==0.0.1
pycares==4.1.2
pycparser==2.21
pymongo==4.0.1
python-dotenv==0.19.2
regex==2022.1.18
requests==2.27.1
urllib3==1.26.8
uvloop==0.16.0
yarg==0.1.9
yarl==1.7.2